Created CFGVisualizeArgs class for abstracting command line arguments
