+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/visualize/CFGVisualizeLauncher.java
@@ -38,7 +38,7 @@ import javax.tools.JavaFileObject;
  * use it.
  */
 public final class CFGVisualizeLauncher {
-
+	
     /** Class cannot be instantiated. */
     private CFGVisualizeLauncher() {
         throw new AssertionError("Class CFGVisualizeLauncher cannot be instantiated.");
@@ -46,80 +46,14 @@ public final class CFGVisualizeLauncher {
 
     /**
      * The main entry point of CFGVisualizeLauncher.
-     *
-     * @param args the passed arguments, see {@link #printUsage()} for the usage
      */
     public static void main(String[] args) {
-        if (args.length == 0) {
-            printUsage();
-            System.exit(1);
-        }
-        String input = args[0];
-        File file = new File(input);
-        if (!file.canRead()) {
-            printError("Cannot read input file: " + file.getAbsolutePath());
-            printUsage();
-            System.exit(1);
-        }
+        CFGVisualizeArgs config = CFGVisualizeArgs.parseArgs(args);
 
-        String method = "test";
-        String clas = "Test";
-        String output = ".";
-        boolean pdf = false;
-        boolean error = false;
-        boolean verbose = false;
-        boolean string = false;
-
-        for (int i = 1; i < args.length; i++) {
-            switch (args[i]) {
-                case "--outputdir":
-                    if (i >= args.length - 1) {
-                        printError("Did not find <outputdir> after --outputdir.");
-                        continue;
-                    }
-                    i++;
-                    output = args[i];
-                    break;
-                case "--pdf":
-                    pdf = true;
-                    break;
-                case "--method":
-                    if (i >= args.length - 1) {
-                        printError("Did not find <name> after --method.");
-                        continue;
-                    }
-                    i++;
-                    method = args[i];
-                    break;
-                case "--class":
-                    if (i >= args.length - 1) {
-                        printError("Did not find <name> after --class.");
-                        continue;
-                    }
-                    i++;
-                    clas = args[i];
-                    break;
-                case "--verbose":
-                    verbose = true;
-                    break;
-                case "--string":
-                    string = true;
-                    break;
-                default:
-                    printError("Unknown command line argument: " + args[i]);
-                    error = true;
-                    break;
-            }
-        }
-
-        if (error) {
-            System.exit(1);
-        }
-
-        if (!string) {
-            generateDOTofCFGWithoutAnalysis(input, output, method, clas, pdf, verbose);
+        if (!config.string) {
+            generateDOTofCFGWithoutAnalysis(config.input, config.output, config.method, config.clas, config.pdf, config.verbose);
         } else {
-            String stringGraph = generateStringOfCFGWithoutAnalysis(input, method, clas, verbose);
+            String stringGraph = generateStringOfCFGWithoutAnalysis(config.input, config.method, config.clas, config.verbose);
             System.out.println(stringGraph);
         }
     }
@@ -354,27 +288,6 @@ public final class CFGVisualizeLauncher {
         Map<String, Object> res = viz.visualize(cfg, cfg.getEntryBlock(), analysis);
         viz.shutdown();
         return res;
-    }
-
-    /** Print usage information. */
-    private static void printUsage() {
-        System.out.println(
-                "Generate the control flow graph of a Java method, represented as a DOT or String"
-                        + " graph.");
-        System.out.println(
-                "Parameters: <inputfile> [--outputdir <outputdir>] [--method <name>] [--class"
-                        + " <name>] [--pdf] [--verbose] [--string]");
-        System.out.println(
-                "    --outputdir: The output directory for the generated files (defaults to '.').");
-        System.out.println(
-                "    --method:    The method to generate the CFG for (defaults to 'test').");
-        System.out.println(
-                "    --class:     The class in which to find the method (defaults to 'Test').");
-        System.out.println("    --pdf:       Also generate the PDF by invoking 'dot'.");
-        System.out.println("    --verbose:   Show the verbose output (defaults to 'false').");
-        System.out.println(
-                "    --string:    Print the string representation of the control flow graph"
-                        + " (defaults to 'false').");
     }
 
     /**
